clean up migration history
packages.download_count has always been bigint

+++ db/migrate/20230531214028_create_packages.rb
@@ -3,7 +3,7 @@ class CreatePackages < ActiveRecord::Migration[7.0]
     create_table :packages do |t|
       t.string :name
       t.references :repository, null: false, foreign_key: true
-      t.integer :download_count
+      t.bigint :download_count
 
       t.timestamps
     end
