+++ db/migrate/20230531214030_create_measurements.rb
@@ -2,7 +2,7 @@ class CreateMeasurements < ActiveRecord::Migration[7.0]
   def change
     create_table :measurements do |t|
       t.references :package, null: false, foreign_key: true
-      t.integer :count
+      t.bigint :count
       t.datetime :measured_at
 
       t.timestamps
